Ft admin register office 164130856

+++ politico/api/v2/office/routes.py
@@ -12,6 +12,13 @@ office_tb = OfficeTable()
 @office.route('/offices', methods=['POST'])
 @token_required
 def create_office(current_user):
+    admin = bool(current_user['is_admin'])
+    if not (admin):
+        print("I am in")
+        return make_response(jsonify({
+                'status': 401,
+                'error': 'Only admin can perform this function'
+            }), 401)
     office_data = request.get_json()
     msg = validate_office_data(office_data)
     if msg != 'ok':
@@ -64,6 +71,13 @@ def get_single_office(current_user, id):
 @office.route('/offices/<int:id>', methods=['PATCH'])
 @token_required
 def update_office(current_user, id):
+    admin = bool(current_user['is_admin'])
+    if not (admin):
+        print("I am in")
+        return make_response(jsonify({
+                'status': 401,
+                'error': 'Only admin can perform this function'
+            }), 401)
     office_data = request.get_json()
     msg = validate_office_data(office_data)
     if msg != 'ok':
@@ -93,6 +107,13 @@ def update_office(current_user, id):
 @office.route('/offices/<int:id>', methods=['DELETE'])
 @token_required
 def delete_office(current_user, id):
+    admin = bool(current_user['is_admin'])
+    if not (admin):
+        print("I am in")
+        return make_response(jsonify({
+                'status': 401,
+                'error': 'Only admin can perform this function'
+            }), 401)
     office = office_tb.get_one_office(id)
     if office:
         if office_tb.delete_office(id):
@@ -129,3 +150,4 @@ def get_office_election_result(current_user, id):
             'status': 200,
             'data': results
         }), 200)
+
